% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpo_data.R
\name{calculate_mpo_long_tt_share}
\alias{calculate_mpo_long_tt_share}
\title{Hour long commute share by Regional Planning Entity}
\usage{
calculate_mpo_long_tt_share(census.yr = 2020, acs.type = "acs5")
}
\arguments{
\item{census.yr}{Four digit integer for Census year for data - defaults to 2020}

\item{acs.type}{String of "acs1" or "acs5" for which ACS type to use - defaults to "acs5"}
}
\value{
tibble of long commute trips by Regional Entity
}
\description{
This function calculates the Share of Commute trips that take at least 1 hour by 27 Regional Planning Entities.
}
\examples{

mpo_hour_commutes <- calculate_mpo_long_tt_share(census.yr=2017)

}
