% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpo_data.R
\name{calculate_mpo_nonsov_share}
\alias{calculate_mpo_nonsov_share}
\title{Non-SOV Mode Share by Regional Planning Entity}
\usage{
calculate_mpo_nonsov_share(census.yr = 2020)
}
\arguments{
\item{census.yr}{Four digit integer for Census year for data - defaults to 2020}
}
\value{
tibble of commute trips and non-sov trips by Regional Entity
}
\description{
This function calculates the Non-SOV Mode Share by 27 Regional Planning Entities.
}
\examples{

mpo_mode_share <- calculate_mpo_nonsov_share(census.yr=2019)

}
