% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpo_data.R
\name{calculate_mpo_ownership_share}
\alias{calculate_mpo_ownership_share}
\title{Home Ownership by Regional Planning Entity}
\usage{
calculate_mpo_ownership_share(census.yr = 2020)
}
\arguments{
\item{census.yr}{Four digit integer for Census year for data - defaults to 2020}
}
\value{
tibble of home ownership shares by Regional Entity
}
\description{
This function calculates the Share of Home Ownership by 27 Regional Planning Entities.
}
\examples{

mpo_home_ownership <- calculate_mpo_ownership_share()

}
