% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpo_data.R
\name{calculate_mpo_people_of_color}
\alias{calculate_mpo_people_of_color}
\title{People of Color by Regional Planning Entity}
\usage{
calculate_mpo_people_of_color(census.yr = 2020, acs.type = "acs5")
}
\arguments{
\item{census.yr}{Four digit integer for Census year for data - defaults to 2020}

\item{acs.type}{String of "acs1" or "acs5" for which ACS type to use - defaults to "acs5"}
}
\value{
tibble of population and people of color by Regional Entity
}
\description{
This function calculates people of color by 27 Regional Planning Entities.
}
\examples{

mpo_people_of_color <- calculate_mpo_people_of_color(census.yr=2018)

}
