% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpo_data.R
\name{calculate_mpo_transit}
\alias{calculate_mpo_transit}
\title{Transit Performance by Regional Planning Entity}
\usage{
calculate_mpo_transit(annual = "no")
}
\arguments{
\item{annual}{yes or no string to determine if the user only wants total annual results - defaults to no}
}
\value{
tibble of transit performance by Regional Entity
}
\description{
This function calculates transit performance from the NTD by 27 Regional Planning Entities.
}
\examples{

mpo_transit <- calculate_mpo_transit(annual='yes')

}
