% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data_trends.R
\name{download_sea_operations_data}
\alias{download_sea_operations_data}
\title{Download SEA Operations Data}
\usage{
download_sea_operations_data(monthyear)
}
\arguments{
\item{monthyear}{Six digit month+year date as string - example "012019"}
}
\value{
downloaded file for current month
}
\description{
This function downloads passenger, air cargo and air operations monthly data for SEA.
The data is based on the Monthly Passenger, Cargo and Operations Summary files.
Data is pulled monthly from "https://www.portseattle.org/page/airport-statistics".
}
\examples{
\dontrun{
download_sea_operations_data(monthyear="042020")
}
}
