% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmer_data.R
\name{get_elmer_table}
\alias{get_elmer_table}
\title{Get a Table from Elmer}
\usage{
get_elmer_table(tbl.nm, srv.nm = "AWS-PROD-SQL\\\\SOCKEYE", db.nm = "Elmer")
}
\arguments{
\item{tbl.nm}{Name of schema and table to pull from the database such as "equity.indicator_facts"}

\item{srv.nm}{Server name, defaults to the current Sockeye instance}

\item{db.nm}{Name of database, defaults to "Elmer"}
}
\value{
tibble of datatable from the central Database
}
\description{
This function pulls a Table from the PSRC Central Database.
}
\examples{

equity.data <- get_elmer_table("equity.indicator_facts")
jurisdictions <- get_elmer_table("Political.jurisdiction_dims")

}
