% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{get_ev_registration_annual}
\alias{get_ev_registration_annual}
\title{Summarize Electric Vehicle Registrations for Calendar Year}
\usage{
get_ev_registration_annual(
  st = c("WA"),
  ct = c("King", "Kitsap", "Pierce", "Snohomish")
)
}
\arguments{
\item{st}{List of States by two digit code to keep - defaults to WA}

\item{ct}{List of County Names to keep - defaults to PSRC counties}
}
\value{
tibble of vehicle registrations by year for selected month
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is originally pulled from https://data.wa.gov/ via the Socrata API
}
\examples{

ev_registration_yearend <- get_ev_registration_annual()

}
