% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{get_ev_registration_monthly}
\alias{get_ev_registration_monthly}
\title{Summarize Electric Vehicle Registrations by County by Month}
\usage{
get_ev_registration_monthly(
  st = c("WA"),
  ct = c("King", "Kitsap", "Pierce", "Snohomish")
)
}
\arguments{
\item{st}{List of States by two digit code to keep - defaults to WA}

\item{ct}{List of County Names to keep - defaults to PSRC counties}
}
\value{
tibble of electric vehicle registrations by month and county
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is originally pulled from https://data.wa.gov/ via the Socrata API
}
\examples{

ev_registration <- get_ev_registration_monthly()

}
