% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{get_ev_registration_ytd}
\alias{get_ev_registration_ytd}
\title{Summarize Electric Vehicle Registrations by County by Year to Date}
\usage{
get_ev_registration_ytd(
  yrs = NULL,
  mos = NULL,
  st = c("WA"),
  ct = c("King", "Kitsap", "Pierce", "Snohomish")
)
}
\arguments{
\item{yrs}{Integer max year if not max from data - default is null}

\item{mos}{Integer month if not max from data - default is null}

\item{st}{List of States by two digit code to keep - defaults to WA}

\item{ct}{List of County Names to keep - defaults to PSRC counties}
}
\value{
tibble of vehicle registrations by year for selected month
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is originally pulled from https://data.wa.gov/ via the Socrata API
}
\examples{

ev_registration_ytd_latest <- get_ev_registration_ytd()
ev_registration_ytd_2021 <- get_ev_registration_ytd(yr=2021)
ev_registration_ytd_april <- get_ev_registration_ytd(mo=4)

}
