% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{get_ev_registrations_county}
\alias{get_ev_registrations_county}
\title{Summarize Electric Vehicle Registrations by County & Year}
\usage{
get_ev_registrations_county(
  ev.url = "X:/DSA/psrc-trends-data/ev-registrations/wa_ev_registrations_public.csv"
)
}
\arguments{
\item{ev.url}{Location of the registration data from Washington State - defaults to X drive}
}
\value{
tibble of electric vehicle registrations by year and county
}
\description{
This function pulls and cleans data from the Washington State Registration Data.
Data is originally pulled from https://www.atlasevhub.com/materials/state-ev-registration-data/#data
}
\examples{

ev_county <- get_ev_registrations_county(ev.url = "C:/Users/chelmann/OneDrive - Puget Sound Regional Council/data-downloads/wa_ev_registrations_public.csv")

}
