% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{get_ofm_postcensal_housing}
\alias{get_ofm_postcensal_housing}
\title{Get OFM Postcensal Housing Data by Jurisdiction}
\usage{
get_ofm_postcensal_housing()
}
\value{
tibble of housing units by jurisdiction
}
\description{
This function pulls Postcensal Housing Data by Jurisdiction.
}
\examples{

ofm.data <- get_ofm_postcensal_housing()

}
