% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{get_ofm_postcensal_pop_change}
\alias{get_ofm_postcensal_pop_change}
\title{Get OFM Postcensal Population Change by Type by County}
\usage{
get_ofm_postcensal_pop_change()
}
\value{
tibble of population change by County
}
\description{
This function pulls Postcensal Population Change Data by County.
}
\examples{

ofm.pop.chg.data <- get_ofm_postcensal_pop_change()

}
