% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{get_population_data}
\alias{get_population_data}
\title{Get OFM Intercensal Population Data by Jurisdiction}
\usage{
get_population_data()
}
\value{
tibble of population by jurisdiction
}
\description{
This function pulls Intercensal Population Data by Jurisdiction.
}
\examples{
\dontrun{
population_data <- get_population_data()}

}
