% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew_data_trends.R
\name{jobs_near_hct}
\alias{jobs_near_hct}
\title{Total Employment Growth Near High Capacity Transit}
\usage{
jobs_near_hct(start_year = 2010)
}
\arguments{
\item{start_year}{The first year of data in the series. Defaults to 2010.}
}
\value{
tibble of total jobs near High Capacity Transit and in the region by calendar year
}
\description{
This function pulls total employment estimates from PSRC's employment database in Elmer.
The jobs estimates are already summarized by year in a table in Elmer.
VISION 2050 Station Area and RGC buffers are stored in ElmerGeo.
}
\examples{
\dontrun{
jobs_hct <- jobs_near_hct()}
 
}
