% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofm_data.R
\name{pop_hsg_near_hct}
\alias{pop_hsg_near_hct}
\title{Population and Housing Unit Growth Near High Capacity Transit}
\usage{
pop_hsg_near_hct()
}
\value{
tibble of population and housing growth near High Capacity Transit by calendar year
}
\description{
This function pulls and cleans data from the SAEP Block data from OFM from Elmer that was parcelized.
}
\examples{
\dontrun{
pop_hsg_hct <- pop_hsg_near_hct()}

 
}
