% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{process_acs_data}
\alias{process_acs_data}
\title{ACS data for PSRC Counties, Cities, Region and MPOs}
\usage{
process_acs_data(
  years = c(2021),
  acs_tbl = "B08301",
  acs_variables = "commute-modes"
)
}
\arguments{
\item{years}{Four digit integer (or vector of integers) for year(s) of analysis}

\item{acs_tbl}{ACS table to be downloaded: defaults to B08301 (Mode to Work)}

\item{acs_variables}{Name of variable look up to use: defaults to commute-modes}
}
\value{
tibble of mode to work data by census year
}
\description{
This function pulls and cleans ACS Data.
Pulls data by year for PSRC Counties, Cities, Region and MPOs
}
\examples{
\dontrun{
mode_to_work <- process_acs_data(years=c(2016, 2021))}

}
