% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data_trends.R
\name{process_faa_data}
\alias{process_faa_data}
\title{Process National Passenger Enplanement Data}
\usage{
process_faa_data(yr1 = 2019, metric = "passengers")
}
\arguments{
\item{yr1}{First four digit calendar year as integer of data to process - defaults to 2019}

\item{metric}{Either "passengers" or "cargo" - defaults to "passengers"}
}
\value{
tibble in long form of FAA Airport Passenger data
}
\description{
This function processes passenger annual data for FAA.
}
\examples{
\dontrun{
national_airport_passengers <- process_faa_data()}

}
