% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data_trends.R
\name{process_ntd_data}
\alias{process_ntd_data}
\title{Download NTD Data Data}
\usage{
process_ntd_data()
}
\value{
tibble of transit related metrics for the region and metro areas
}
\description{
This function cleans data from the Monthly NTD Raw Data Release.
Data is downloaded from https://www.transit.dot.gov/ntd/data-product/monthly-module-raw-data-release
}
\examples{
\dontrun{
transit_data <- process_ntd_data()}

}
