% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data_trends.R
\name{process_ntd_monthly_data}
\alias{process_ntd_monthly_data}
\title{Process NTD Monthly Transit Data}
\usage{
process_ntd_monthly_data()
}
\value{
tibble in long form of monthly Boardings, Revenue-Miles and Revenue-Hours by Agency and Region
}
\description{
This function processes Transit Agency monthly data from the Nation Transit Database.
Data is pulled monthly from "https://www.transit.dot.gov/sites/fta.dot.gov/files/".
}
\examples{

ntd.data <- process_ntd_monthly_data()

}
