% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data_trends.R
\name{process_ntd_uza_data}
\alias{process_ntd_uza_data}
\title{Summarize Transit Data by Urbanized Area}
\usage{
process_ntd_uza_data(yr, pop.limit = 1e+06, census.yr = "2020")
}
\arguments{
\item{yr}{Four digit calendar year as string for Summary year}

\item{pop.limit}{Population threshold to use to filter down Urbanized Areas - defaults to 1,000,000}

\item{census.yr}{Four digit calendar year as string for Census Year for UZA population data - defaults to "2020"}
}
\value{
tibble in long form of Year to Date (or annual if it is a compelte year) Boardings and Revenue-Hours by Urbanized Area
}
\description{
This function processes Transit Agency monthly data from the Nation Transit Database and aggregate to the Urbanized Areas.
Data is pulled monthly from "https://www.transit.dot.gov/sites/fta.dot.gov/files/".
}
\examples{

uza.data <- process_ntd_uza_data(yr="2022", census.yr="2020")

}
