% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd_data_trends.R
\name{process_ntd_year_to_date_data}
\alias{process_ntd_year_to_date_data}
\title{Summarize Year to Date Transit Data by Mode and Operator}
\usage{
process_ntd_year_to_date_data(c.yr, c.mo = NULL)
}
\arguments{
\item{c.yr}{Four digit calendar year as string for last year to analyze}

\item{c.mo}{Month for analysis in year to date calculations if the latest month isn't desired - defaults to NULL}
}
\value{
tibble in long form of Year to Date (or annual if it is a complete year) of Transit Data by Mode and Operator
}
\description{
This function processes Transit Agency monthly data from the Nation Transit Database and aggregate to year to date.
Data is pulled monthly from "https://www.transit.dot.gov/sites/fta.dot.gov/files/".
}
\examples{

ntd.ytd <- process_ntd_year_to_date_data(c.yr="2022")

}
