% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew_data_trends.R
\name{process_qcew_monthly_msa}
\alias{process_qcew_monthly_msa}
\title{Process QCEW Monthly MSA Data}
\usage{
process_qcew_monthly_msa(yr, mo)
}
\arguments{
\item{yr}{Current four digit calendar year as integer}

\item{mo}{Current month as integer, no leading zeros}
}
\value{
tibble in long form of QCEW Monthly Employment data by MSA+
}
\description{
This function processes MSA monthly data from QCEW.
Data is pulled monthly from "https://media.esd.wa.gov/esdwa/Default/ESDWAGOV/labor-market-info/Libraries/Economic-reports/Washington-employment-estimates/".
}
\examples{
\dontrun{
jobs_data <- process_qcew_monthly_msa(yr=2023, mo=12)}

}
