% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data_trends.R
\name{process_sea_operations_data}
\alias{process_sea_operations_data}
\title{Process SEA Operations Data}
\usage{
process_sea_operations_data(c.yr, c.mo, f.yr = 2019)
}
\arguments{
\item{c.yr}{Current four digit calendar year as integer}

\item{c.mo}{Current month as integer, no leading zeros}

\item{f.yr}{First four digit calendar year as integer of data to process - defaults to 201*}
}
\value{
tibble in long form of SeaTac Airport operations data
}
\description{
This function process passenger, air cargo and air operations monthly data for SEA.
The data is based on the Monthly Passenger, Cargo and Operations Summary files.
Monthly data is stored on PSRC servers at "X:/DSA/psrc-trends-data/monthly-seatac".
Data is pulled monthly from "https://www.portseattle.org/page/airport-statistics".
}
\examples{

airport.operations <- process_sea_operations_data(c.yr=2022, c.mo=4)

}
