% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{process_vehicle_registration_data}
\alias{process_vehicle_registration_data}
\title{Vehicle Registration data by various geographies}
\usage{
process_vehicle_registration_data(
  dol_registration_file = "X:/DSA/rtp-dashboard/DOL/Vehicle_Title_Transactions.csv"
)
}
\arguments{
\item{dol_registration_file}{path to downloaded registration data file}
}
\value{
tibble vehicle registrations by electrification level for county, region and census tracts
}
\description{
This function summarizes vehicle registration data by electrification level 
The data is pre-downloaded due to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
vehicle_data <- process_vehicle_registration_data()}

}
