% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew_data_trends.R
\name{processed_covered_employment}
\alias{processed_covered_employment}
\title{Total Observed and Forecast Employment Growth}
\usage{
processed_covered_employment(forecast_base_yr = 2018, first_year = 2010)
}
\arguments{
\item{forecast_base_yr}{Four digit integer for Base Year in Forecast Data - defaults to 2018}

\item{first_year}{Four digit integer for first year of data to use from Elmer - defaults to 2010}
}
\value{
tibble of total jobs in the region by calendar year
}
\description{
This function pulls total employment estimates from PSRC's employment database in Elmer and the Macroforecast.
The jobs estimates are already summarized by year in a table in Elmer.
}
\examples{
\dontrun{
jobs <- processed_covered_employment()}
 
}
