% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrctrends_details.R
\docType{package}
\name{psrctrends}
\alias{psrctrends}
\alias{psrctrends-package}
\title{psrctrends: A package for easy data processing for common Trends in the Puget Sound Region.}
\description{
Various functions to process trend data in the Central Puget Sound Region.
This package is currently under development and will be updated frequently over the last half of 2022.
}
\author{
Craig Helmann \email{chelmann@psrc.org}
}
