% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data_trends.R
\name{summarize_airport_data}
\alias{summarize_airport_data}
\title{Summarize Airport Data}
\usage{
summarize_airport_data(yr1 = 2019)
}
\arguments{
\item{yr1}{First four digit calendar year as integer of data to process - defaults to 2019}
}
\value{
tibble in long form of SeaTac Airport operations data
}
\description{
This function summarizes passenger, air cargo and air operations monthly data for SEA.
The data is based on the Monthly Passenger, Cargo and Operations Summary files.
}
\examples{
\dontrun{
airport_data <- summarize_airport_data(yr1=2022)}

}
