% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{summarize_census_data}
\alias{summarize_census_data}
\title{Summarize Census Data}
\usage{
summarize_census_data(years = c(2021))
}
\arguments{
\item{years}{Four digit integer (or vector of integers) for year(s) of analysis}
}
\value{
tibble in long form of ACS data
}
\description{
This function summarizes ACS Census Data fro Cities, Counties, Region and MPOs
}
\examples{
\dontrun{
census_data <- summarize_census_data(years=c(2022))}

}
