% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcew_data_trends.R
\name{summarize_jobs_data}
\alias{summarize_jobs_data}
\title{Summarize Jobs Data}
\usage{
summarize_jobs_data(yr, mo)
}
\arguments{
\item{yr}{Current four digit calendar year as integer}

\item{mo}{Current month as integer, no leading zeros}
}
\value{
tibble in long form of Jobs data
}
\description{
This function summarizes Covered and Wage & Slary Jobs Data
}
\examples{
\dontrun{
jobs_data <- summarize_jobs_data(yr=2023, mo=12)}

}
