% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electric_vehicles.R
\name{vehicle_manufacturers}
\alias{vehicle_manufacturers}
\title{Electric Vehicle Makers by Vehicle Registration}
\usage{
vehicle_manufacturers(
  data_file = "X:/DSA/rtp-dashboard/DOL/Vehicle_Title_Transactions.csv",
  engine_type = "Battery Electric Vehicle"
)
}
\arguments{
\item{data_file}{path to downloaded data files https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x}

\item{engine_type}{type of vehicle to summarize make for - defaults to Battery Electric Vehicle}
}
\value{
tibble of electric vehicle manufacturers registered in the region by year to date
}
\description{
This cleans data from the Washington State Registration Data vehicle registrations on open data portal.
The data is pre-downloaded due to size from: https://data.wa.gov/Transportation/Title-Transactions-by-Month/u4cd-bc3x
}
\examples{
\dontrun{
ev_manufacturers <- vehicle_manufacturers(engine_type="Battery Electric Vehicle")}

}
